Move blocks lists
